apply the same double precision change to PF_normalize() as was done to PF_vlen() in r1554.
git-svn-id: https://svn.code.sf.net/p/quakespasm/code/trunk@1578 af15c1b1-3010-417e-b628-4374ebc0bcbd

+++ Quake/pr_cmds.c
@@ -397,11 +397,11 @@ static void PF_normalize (void)
 {
 	float	*value1;
 	vec3_t	newvalue;
-	float	new_temp;
+	double	new_temp;
 
 	value1 = G_VECTOR(OFS_PARM0);
 
-	new_temp = value1[0] * value1[0] + value1[1] * value1[1] + value1[2]*value1[2];
+	new_temp = (double)value1[0] * value1[0] + (double)value1[1] * value1[1] + (double)value1[2]*value1[2];
 	new_temp = sqrt (new_temp);
 
 	if (new_temp == 0)
